transport limited cleanup, tests
TL module now has elementary test for its mono directional routing
version (polydirectional remains untested).

Cleaned up unused, older code iterations.
