+++ landlab/components/transport_limited_fluvial/tests/stream_power_params_ideal.txt
@@ -1,0 +1,40 @@
+#These parameters produce ideal output, for “annual-ish” storm ballpark conditions.
+#Up k_Q by an order of magnitude (to 2.6e-6) to simulate constant ~50 yr flood 
+#conditions, and to induce much more prominent stream concavity.
+nrows:
+50
+ncols:
+50
+dx:
+1000.
+leftmost_elevation:
+0.
+initial_slope:
+0.
+dt:
+10000.
+total_time:
+100000.
+uplift_rate:
+0.0005
+#time is years, ALL linear dimensions are meters!
+
+#these are the parameters for the stream power module:
+b_sp:
+0.5
+c_sp:
+1.
+Qc:
+MPM
+k_Q:
+2.5e-7
+k_w:
+2.5
+mannings_n:
+0.05
+Dchar:
+0.1
+threshold_shields:
+0.06
+return_stream_properties:
+True
