+++ landlab/components/transport_limited_fluvial/tl_fluvial_monodirectional.py
@@ -1,11 +1,16 @@
 from __future__ import print_function
 
 import numpy as np
+import inspect
 from landlab import ModelParameterDictionary, CLOSED_BOUNDARY
+from landlab import RasterModelGrid
 from time import sleep
+from landlab.utils import structured_grid as sgrid
+import pylab
 
 from landlab.core.model_parameter_dictionary import MissingKeyError
 from landlab.field.scalar_data_fields import FieldError
+from landlab.grid.base import BAD_INDEX_VALUE
 
 
 class TransportLimitedEroder(object):
@@ -17,9 +22,7 @@ class TransportLimitedEroder(object):
     The module can in principle take multiple transport laws, but at the moment
     only Meyer-Peter Muller (MPM) is implemented.
 
-    There is as yet no explicit stabilization check on the timestep. If your
-    run destabilizes, try reducing dt.
-    See, e.g., ./examples/simple_sp_driver.py
+    This module is defined to interface neatly with the flow_routing module.
 
     Assumes grid does not deform or change during run.
 
@@ -28,16 +31,12 @@ class TransportLimitedEroder(object):
     and the uplift rate...!)
 
     DEJH Sept 14.
-    This component *should* run on any grid, but untested.
+    (Only tested for raster grid so far)
     """
 
     def __init__(self, grid, params):
         self.initialize(grid, params)
 
-#This draws attention to a potential problem. It will be easy to have modules update z, but because noone "owns" the data, to forget to also update dz/dx...
-#How about a built in grid utility that updates "derived" data (i.e., using only grid fns, e.g., slope, curvature) at the end of any given tstep loop?
-#Or an explicit flagging system for all variables in the modelfield indicating if they have been updated this timestep. (Currently implemented)
-#Or wipe the existance of any derived grid data at the end of a timestep entirely, so modules find they don't have it next timestep.
 
     def initialize(self, grid, params_file):
         '''
@@ -68,14 +67,13 @@ class TransportLimitedEroder(object):
                 from the Manning's equation, respectively. These are
                 needed to allow calculation of shear stresses and hence carrying
                 capacities from the local slope and drainage area alone.
-                The equation for depth used to derive shear stress and hence
-                carrying capacity contains a prefactor:
-                    mannings_n*(k_Q**(1-b)/K_w)**0.6
-                (so shear = fluid_density*g*depth_equation_prefactor*A**(0.6*c*(1-b)*S**0.7 !)
-                Don't know what to set these values to? k_w=0.002, k_Q=1.e-9,
-                mannings_n=0.03 give vaguely plausible numbers (e.g., for a
-                drainage area ~400km2, like Boulder Creek at Boulder,
-                => depth~2.5m, width~35m, shear stress ~O(1000Pa)).
+                Don't know what to set these values to? k_w=2.5, k_Q=2.5e-7,
+                mannings_n=0.05 give vaguely plausible numbers with b=0.5,
+                c = 1.(e.g., for a drainage area ~350km2, like Boulder Creek
+                at Boulder, => depth~1.3m, width~23m, shear stress ~O(200Pa)
+                for an "annual-ish" flood). [If you want to continue playing
+                with calibration, the ?50yr return time 2013 floods produced
+                depths ~2.3m with Q~200m3/s]
             *Dchar -> float.  The characteristic grain diameter in meters
                 (==D50 in most cases) used to calculate Shields numbers
                 in the channel. If you want to define Dchar values at each node,
@@ -95,7 +93,8 @@ class TransportLimitedEroder(object):
                     threshold_shields=0.15*S**0.25
             *slope_sensitive_threshold -> bool, defaults to 'False'.
                 If true, threshold_shields is set according to the Lamb
-                equation. An exception will be raised if threshold_shields is
+                equation,
+                An exception will be raised if threshold_shields is
                 also set.
             *dt -> +ve float. If set, this is the fixed timestep for this
                 component. Can be overridden easily as a parameter in erode().
@@ -103,17 +102,24 @@ class TransportLimitedEroder(object):
             *use_W -> Bool; if True, component will look for node-centered data
                 describing channel width in grid.at_node['channel_width'], and
                 use it to implement incision ~ stream power per unit width.
-                Defaults to False.
+                Defaults to False. NOT YET IMPLEMENTED
             *use_Q -> Bool. Overrides the basin hydrology relation, using an
                 local water discharge value assumed already calculated and
-                stored in grid.at_node['discharge'].
+                stored in grid.at_node['discharge']. NOT YET IMPLEMENTED
             *C_MPM -> float. Defaults to 1. Allows tuning of the MPM prefactor,
                 which is calculated as
                     Qc = 8.*C_MPM*(taustar - taustarcrit)**1.5
                 In almost all cases, tuning depth_equation_prefactor' is
                 preferred to tuning this parameter.
+            *return_stream_properties -> bool (default False).
+                If True, this component will save the calculations for
+                'channel_width', 'channel_depth', and 'channel_discharge' in
+                those grid fields. (Requires some
+                additional math, so is suppressed for speed by default).
 
         '''
+        #this is the fraction we allow any given slope in the grid to evolve by in one go (suppresses numerical instabilities)
+        self.fraction_gradient_change = 0.25
         self.grid = grid
         self.link_S_with_trailing_blank = np.zeros(grid.number_of_links+1) #needs to be filled with values in execution
         self.count_active_links = np.zeros_like(self.link_S_with_trailing_blank, dtype=int)
@@ -135,9 +141,7 @@ class TransportLimitedEroder(object):
             self.fluid_density = inputs.read_float('fluid_density')
         except MissingKeyError:
             self.fluid_density = 1000.
-        self.relative_weight = (self.sed_density-self.fluid_density)/self.fluid_density*self.g #to accelerate MPM calcs
-        self.excess_SG = (self.sed_density-self.fluid_density)/self.fluid_density
-        self.rho_g = self.fluid_density*self.g
+        self.rho_g = self.fluid_density * self.g
 
         try:
             self.Qc = inputs.read_string('Qc')
@@ -148,6 +152,10 @@ class TransportLimitedEroder(object):
                 self.calc_cap_flag = True
             else:
                 self.calc_cap_flag = False
+        try:
+            self.return_ch_props = inputs.read_bool('return_stream_properties')
+        except MissingKeyError:
+            self.return_ch_props = False
 
         try:
             self.lamb_flag = inputs.read_bool('slope_sensitive_threshold')
@@ -174,6 +182,10 @@ class TransportLimitedEroder(object):
             self.use_Q = inputs.read_bool('use_Q')
         except MissingKeyError:
             self.use_Q = False
+        try:
+            self.return_capacity = inputs.read_bool('return_capacity')
+        except MissingKeyError:
+            self.return_capacity = False
 
         try:
             self._b = inputs.read_float('b_sp')
@@ -199,139 +211,67 @@ class TransportLimitedEroder(object):
         #assume Manning's equation to set the power on A for shear stress:
         self.shear_area_power = 0.6*self._c*(1.-self._b)
 
-        try:
-            self.k_Q = inputs.read_float('k_Q')
-        except MissingKeyError:
-            self.depth_prefactor = self.rho_g*inputs.read_float('depth_equation_prefactor')
-        else:
-            self.k_w = inputs.read_float('k_w')
-            mannings_n = inputs.read_float('mannings_n')
-            if mannings_n<0. or mannings_n>0.2:
-                print("***STOP. LOOK. THINK. You appear to have set Manning's n outside it's typical range. Did you mean it? Proceeding...***")
-                sleep(2)
-            self.depth_prefactor = self.rho_g*mannings_n*(self.k_Q**(1.-self._b)/self.k_w)**0.6
-            ##Note the depth_prefactor we store already holds rho*g
+        self.k_Q = inputs.read_float('k_Q')
+        self.k_w = inputs.read_float('k_w')
+        mannings_n = inputs.read_float('mannings_n')
+        self.mannings_n = mannings_n
+        if mannings_n<0. or mannings_n>0.2:
+            print("***STOP. LOOK. THINK. You appear to have set Manning's n outside its typical range. Did you mean it? Proceeding...***")
+            sleep(2)
 
         try:
             self.C_MPM = inputs.read_float('C_MPM')
         except MissingKeyError:
             self.C_MPM = 1.
-        try:
-            self.shields_prefactor = 1./((self.sed_density-self.fluid_density)*self.g*self.Dchar_in)
-            self.MPM_prefactor = 8.*self.C_MPM*np.sqrt(self.relative_weight*self.Dchar_in*self.Dchar_in*self.Dchar_in)
-            self.MPM_prefactor_alt = 4.*self.g**(-2./3.)/self.excess_SG/self.fluid_density/self.sed_density
+        self.diffusivity_power_on_A = 0.9*self._c*(1.-self._b) #i.e., q/D**(1/6)
+
+        #new for v3:
+        #set thresh in shear stress if poss at this stage:
+        try: #fails if no Dchar provided, or shields crit is being set dynamically from slope
+            self.thresh = self.shields_crit*(self.sed_density-self.fluid_density)*self.g*self.Dchar_in
         except AttributeError:
-            #have to set these manually as needed
-            self.shields_prefactor_noD = 1./((self.sed_sensity-self.fluid_density)*self.g)
-
-        self.cell_areas = np.empty(grid.number_of_nodes)
-        self.cell_areas.fill(np.mean(grid.area_of_cell))
-        self.cell_areas[grid.node_at_cell] = grid.area_of_cell
-
-
-    def sed_capacity_equation(self, grid, shields_stress, slopes_at_nodes=None, areas_at_node=None):
-        """
-        slopes_at_nodes must be provided as nnodes array if slope_sensitive_threshold.
-        This is a volume flux.
-        At the moment, this only supports MPM for sed flux capacity calculation,
-        but is set up to allow easy additions later if needed.
-        """
-        if self.lamb_flag:
-            thresh = 0.15*slopes_at_nodes**0.25
-        else:
-            thresh = self.shields_crit
-        if self.Qc == 'MPM':
-            print("in capacity:")
-            print('prefactor ', self.MPM_prefactor)
-            print('max Shields ', np.amax(shields_stress[self.grid.core_nodes]))
-            print('thresh ', np.amax(thresh))
             try:
-                capacity = self.MPM_prefactor*((shields_stress-thresh).clip(0.))**1.5
-            except AttributeError:
-                capacity = 8.*self.C_MPM*np.sqrt(self.relative_weight*self.Dchar*self.Dchar*self.Dchar)*((shields_stress-thresh).clip(0.))**1.5
-        elif type(self.Qc)==str:
-            try:
-                capacity = grid.at_node[self.Qc]
-            except:
-                raise TypeError('sed_flux_dep_incision does not know how to set sed transport capacity from the provided input...')
-        elif type(self.Qc)==np.ndarray:
-            capacity=self.Qc
+                self.shields_prefactor_to_shear = (self.sed_density-self.fluid_density)*self.g*self.Dchar_in
+            except AttributeError: #no Dchar
+                self.shields_prefactor_to_shear_noDchar = (self.sed_density-self.fluid_density)*self.g
+        twothirds = 2./3.
+        self.Qs_prefactor = 4.*self.C_MPM**twothirds*self.fluid_density**twothirds/(self.sed_density-self.fluid_density)**twothirds*self.g**(twothirds/2.)*mannings_n**0.6*self.k_w**(1./15.)*self.k_Q**(0.6+self._b/15.)/self.sed_density**twothirds
+        self.Qs_thresh_prefactor = 4.*(self.C_MPM*self.k_w*self.k_Q**self._b/self.fluid_density**0.5/(self.sed_density-self.fluid_density)/self.g/self.sed_density)**twothirds
+        #both these are divided by sed density to give a vol flux
+        self.Qs_power_onA = self._c*(0.6+self._b/15.)
+        self.Qs_power_onAthresh = twothirds*self._b*self._c
+
+        if RasterModelGrid in inspect.getmro(grid.__class__):
+            self.cell_areas = grid.dx * grid.dy
         else:
-            raise TypeError('sed_flux_dep_incision does not know how to set sed transport capacity from the provided inputs...')
-        capacity *= self.k_w*areas_at_node**self._b*31557600.
-        return capacity #returned capacity is volume flux/yr
+            self.cell_areas = np.empty(grid.number_of_nodes)
+            self.cell_areas.fill(np.mean(grid.area_of_cell))
+            self.cell_areas[grid.node_at_cell] = grid.area_of_cell
+        self.bad_neighbor_mask = np.equal(grid.get_active_neighbors_at_node(bad_index=-1),-1)
 
-    def sed_capacity_equation_alt_method(self, grid, shear_stress, slopes_at_nodes=None, areas_at_node=None):
-        """
-        slopes_at_nodes must be provided as nnodes array if slope_sensitive_threshold.
-        This is a volume flux.
-        At the moment, this only supports MPM for sed flux capacity calculation,
-        but is set up to allow easy additions later if needed.
-        """
-        if self.lamb_flag:
-            thresh = 0.15*slopes_at_nodes**0.25
-        else:
-            thresh = self.shields_crit
-        if self.Qc == 'MPM':
-            capacity = self.MPM_prefactor_alt*((shear_stress-self.excess_SG*self.Dchar*self.rho_g*thresh).clip(0.))
-        elif type(self.Qc)==str:
-            try:
-                capacity = grid.at_node[self.Qc]
-            except:
-                raise TypeError('sed_flux_dep_incision does not know how to set sed transport capacity from the provided input...')
-        elif type(self.Qc)==np.ndarray:
-            capacity=self.Qc
-        else:
-            raise TypeError('sed_flux_dep_incision does not know how to set sed transport capacity from the provided inputs...')
-        capacity *= self.k_w*areas_at_node**self._b*31557600.
-        return capacity #returned capacity is volume flux/yr
+        self.routing_code = """
+            double sed_flux_into_this_node;
+            double sed_flux_out_of_this_node;
+            double flux_excess;
+            for (int i=len_s_in; i>0; i--) {
+                sed_flux_into_this_node = sed_into_node[i];
+                sed_flux_out_of_this_node = transport_capacities[i];
+                flux_excess = sed_flux_into_this_node - sed_flux_out_of_this_node;
+                dz[i] = flux_excess/cell_areas*dt_this_step;
+                sed_into_node[flow_receiver[i]] += sed_flux_out_of_this_node;
+            }
+            """
 
 
-    def weave_iter_sed_trp_balance(self, grid, capacity, r, s, dt):
-        '''
-        Needs to be passed at node values for 'flow_receiver' and
-        'upstream_node_order' (i.e., r and s). These are most likely the output
-        values from route_flow_dn.
-        Returns dz/dt, the rate of elevation change.
-        '''
-        code =  """
-                int donor;
-                int rcvr;
-                for (int i = num_pts-1; i > -1; i--) {
-                    donor = s[i];
-                    rcvr = r[donor];
-                    dz[donor] = (sed_into_cell[donor] - capacity[donor]*dt)/cell_areas[donor];
-                    if (donor != rcvr) {
-                        sed_into_cell[rcvr] += capacity[donor];
-                    }
-                }
-                """
-        num_pts = len(s)
-        cell_areas = self.cell_areas #nnodes long
-        sed_into_cell = np.zeros(capacity.size, dtype=np.float64) #this is, more accurately, a sed *flux* in
-        dz = np.zeros_like(sed_into_cell)
-        #print type(cell_areas), type(s), type(r), type(capacity), type(sed_into_cell)
-
-        for donor in s[::-1]:
-            rcvr = r[donor]
-            #by the time we get to donor, no more sed will be added; it's "done":
-            ########where should the next line be? Can an internally drained node erode down/deposit?
-            dz[donor] = (sed_into_cell[donor]-capacity[rcvr]*dt)/cell_areas[donor] #if capacity is greater than sed_in, it will erode (-ve value here)
-            #we look downstream to ensure continuity with the BCs at grid margins
-            #print "dz: ", dz[donor]
-            #capacity's worth of sed leaves donor & goes to rcvr
-            if donor != rcvr:
-                #print "saved", donor, rcvr
-                #print "dz: ", dz[donor]
-                sed_into_cell[rcvr] += capacity[donor]
-
-        return dz
-
-
-    def erode(self, grid, dt, node_drainage_areas='drainage_area',
-                slopes_at_nodes=None, link_slopes=None, link_node_mapping='links_to_flow_receiver',
-                receiver='flow_receiver', upstream_order='upstream_node_order',
-                slopes_from_elevs=None, W_if_used=None, Q_if_used=None,
+    def erode(self, grid, dt=None, node_elevs='topographic__elevation',
+                node_drainage_areas='drainage_area',
+                node_receiving_flow='flow_receiver',
+                node_order_upstream='upstream_node_order',
+                node_slope='topographic__steepest_slope',
+                steepest_link='links_to_flow_receiver',
+                runoff_rate_if_used=None,
+                #W_if_used=None, Q_if_used=None,
+                stability_condition='loose',
                 Dchar_if_used=None, io=None):
 
         """
@@ -340,14 +280,21 @@ class TransportLimitedEroder(object):
 
         *grid* & *dt* are the grid object and timestep (float) respectively.
 
-        *Node_drainage_areas* tells the component where to look for the drainage
+        *node_elevs* tells the component where to look for the node elevations.
+        Pass another string to override which grid field the component looks
+        at, or pass a nnodes-long array of elevation values directly instead.
+
+        *node_drainage_areas* tells the component where to look for the drainage
         area values. Change to another string to override which grid field the
         component looks at, or pass a nnodes-long array of drainage areas values
         directly instead.
 
-        If you already have slopes defined at nodes on the grid, pass them to
-        the component with *slopes_at_nodes*. The same syntax is expected:
-        string gives a name in the grid fields, an array gives values direct.
+        *node_receiving flow* tells the component where to look for the node
+        ids which receive flow from each node. This is an output from the
+        flow_routing module.
+        Change to another string to override which grid field the
+        component looks at, or pass a nnodes-long array of IDs
+        directly instead.
 
         Alternatively, set *link_slopes* (and *link_node_mapping*) if this data
         is only available at links. 'planet_surface__derivative_of_elevation'
@@ -376,27 +323,61 @@ class TransportLimitedEroder(object):
         slope. This is primarily for speed, but may be a good idea to modify
         later.
 
+        *runoff_rate_if_used* is the runoff rate in m/yr, if used (take care
+        with the units...!). Either a float, or an nnodes-long array
+        (NB: array functionality is untested).
+        If specified, becomes a multiplicative modifier on k_Q, above.
+        Ensure you adjust the precipitation time series so that the
+        flood series you get makes sense! If not set, the precip rate
+        is assumed to be rolled into the k_Q term already.
+
         *W_if_used* and *Q_if_used* must be provided if you set use_W and use_Q
         respectively in the component initialization. They can be either field
         names or nnodes arrays as in the other cases.
+
+        *stability_condition* controls how we limit the internal timestep of
+        the model to improve solution stability. 'loose' (default) employs a
+        condition that prevents changes in the drainage structure of the
+        existing channel network - slopes may not reverse - and this should
+        prove adequate for most uses. 'tight' uses a considerably stricter Lax/
+        Von Neumann criterion, but will be considerably slower.
 
         *Dchar_if_used* must be set as a grid field string or nnoodes-long array
         if 'Dchar' as a float was not provided in the input file. (If it was,
         this will be overridden).
 
-        RETURNS XXXXXX
+        SETS: (as fields on the grid)
+        ***Note the time units are SECONDS in these fields***
+        - 'topographic__elevation' (m), the elevations (or your name)
+        - 'fluvial_sediment_transport_capacity' (m**3/s), the volumetric
+            transport capacities at each node
+        - 'fluvial_sediment_flux_into_node' (m**3/s), the total volumetric sed
+            flux entering each node
+        If your stability_condition was 'tight':
+        - 'effective_fluvial_diffusivity'
+        If return_stream_properties was True:
+        - 'channel_width' (m)
+        - 'channel_depth' (m)
+        - 'channel_discharge' (m**3/s)
+        - 'channel_bed_shear_stress' (Pa)
+        - The number of internal loops required per supplied time step is
+            stored as a property of the class instance, self.iterations_in_dt.
+            This can be useful for tracking computational load imposed by this
+            component.
+
+        RETURNS:
+        grid, elevations
+
         """
 
-        try:
-            r_in = grid.at_node[receiver]
-        except FieldError:
-            r_in = receiver
-            assert r_in.size == grid.number_of_nodes
-        try:
-            s_in = grid.at_node[upstream_order]
-        except FieldError:
-            s_in = upstream_order
-            assert s_in.size == grid.number_of_nodes
+        if runoff_rate_if_used != None:
+            runoff_rate = runoff_rate_if_used
+            assert type(runoff_rate) in (int, float, np.ndarray)
+        else:
+            runoff_rate = 1.
+
+        if dt==None:
+            dt = self.tstep
         try:
             self.Dchar=self.Dchar_in
         except AttributeError:
@@ -406,88 +387,167 @@ class TransportLimitedEroder(object):
                 assert type(Dchar_if_used)==np.ndarray
                 self.Dchar=Dchar_if_used
 
-        #Perform check on whether we use grid or direct fed data:
-        if slopes_at_nodes==None:
-            if slopes_from_elevs:
-                if slopes_from_elevs == True:
-                    node_z = grid.at_node['topographic__elevation']
-                elif type(slopes_from_elevs) == str:
-                    node_z = grid.at_node[slopes_from_elevs]
-                else:
-                    node_z = slopes_from_elevs
-                S_links = (node_z[grid.node_at_link_tail]-node_z[grid.node_at_link_head])/grid.link_length
-            else:
-                if link_slopes:
-                    if type(link_slopes)==str:
-                        S_links = grid.at_link[link_slopes]
-                    else:
-                        S_links = link_slopes
-                else:
-                    S_links = grid.at_link['planet_surface__derivative_of_elevation']
-
-            #put the slopes onto the nodes
-            try:
-                self.slopes = S_links[grid.at_node[link_node_mapping]]
-            except FieldError:
-                try:
-                    self.slopes = S_links[link_node_mapping]
-                except IndexError:
-                    #need to do the mapping on the fly.
-                    #we're going to use the max slope (i.e., + or -) of *all* adjacent nodes.
-                    #This isn't ideal. It should probably just be the outs...
-                    #i.e., np.amax(self.link_S_with_trailing_blank[grid.node_outlinks] AND -self.link_S_with_trailing_blank[grid.node_inlinks])
-                    self.link_S_with_trailing_blank[:-1] = S_links
-                    self.slopes = np.amax(np.fabs(self.link_S_with_trailing_blank[grid.node_links]),axis=0)
+        if type(node_elevs)==str:
+            node_z = grid.at_node[node_elevs]
         else:
-            try:
-                self.slopes = grid.at_node[slopes_at_nodes]
-            except FieldError:
-                self.slopes = slopes_at_nodes
+            node_z = node_elevs
 
-        #print 'Max slope ', np.amax(self.slopes[self.grid.core_nodes])
-        #print 'Mean slope ', np.mean(self.slopes[self.grid.core_nodes])
         if type(node_drainage_areas)==str:
             node_A = grid.at_node[node_drainage_areas]
         else:
             node_A = node_drainage_areas
 
-        #print 'Area ', np.amax(node_A)
-        #calc Shields
-######need to add ability to do this with W or Q scaling specified
-        shear_stress = self.depth_prefactor*node_A**self.shear_area_power*self.slopes**0.7
+        if type(node_receiving_flow)==str:
+            flow_receiver = grid.at_node[node_receiving_flow]
+        else:
+            flow_receiver = node_receiving_flow
+
+        #new V3:
+        if type(node_order_upstream)==str:
+            s_in = grid.at_node[node_order_upstream]
+        else:
+            s_in = node_order_upstream
+
+        if type(node_slope)==str:
+            node_S = grid.at_node[node_slope]
+        else:
+            node_S = node_slope
+
+        if self.lamb_flag:
+            variable_shields_crit = 0.15*node_S**0.25
+            try:
+                variable_thresh = variable_shields_crit*self.shields_prefactor_to_shear
+            except AttributeError:
+                variable_thresh = variable_shields_crit*self.shields_prefactor_to_shear_noDchar*self.Dchar
+
+
+        if type(steepest_link)==str:
+            link_length = np.empty(grid.number_of_nodes,dtype=float)
+            link_length.fill(np.nan)
+            draining_nodes = np.not_equal(grid.at_node[steepest_link], BAD_INDEX_VALUE)
+            core_draining_nodes = np.intersect1d(np.where(draining_nodes)[0], grid.core_nodes, assume_unique=True)
+            link_length[core_draining_nodes] = grid.link_length[grid.at_node[steepest_link][core_draining_nodes]]
+            #link_length=grid.dx
+        else:
+            link_length = grid.link_length[steepest_link]
+        square_link_length = np.square(link_length) #nans propagate forward
+
         try:
-            shields_stress = self.shields_prefactor*shear_stress
+            transport_capacities_thresh = self.thresh*self.Qs_thresh_prefactor*runoff_rate**(0.66667*self._b)*node_A**self.Qs_power_onAthresh
         except AttributeError:
-            shields_stress = self.shields_prefactor_noD/self.Dchar*shear_stress
-        #print 'max_shear ', np.amax(shear_stress[self.grid.core_nodes])
-        #capacity = self.sed_capacity_equation(grid, shields_stress, self.slopes, node_A)
-        capacity = self.sed_capacity_equation_alt_method(grid, shear_stress, self.slopes, node_A)
-        #print capacity
-        #need to strip closed nodes from s:
-        #...but in fact we use the mask on capacity, not a BC check, as some open nodes can still have no slopes defined (so are functionally closed) - i.e., the corners
-        try:
-            s_in = np.delete(s_in, np.where(np.in1d(s_in,np.where(capacity.mask),assume_unique=True))) #yuck! #this isn't very memory efficient...
-        except AttributeError: #provided S array might not be masked...
-            s_in = np.delete(s_in, np.where(self.grid.is_boundary(s_in, boundary_flag=4)))
-        #print "r: ", r_in
-        #print "s: ", s_in
-        dz = self.weave_iter_sed_trp_balance(grid, capacity, r_in, s_in, dt)
-        #print 'max inc rate ', np.amax(np.fabs(dz[self.grid.core_nodes]))
+            transport_capacities_thresh = variable_thresh*self.Qs_thresh_prefactor*runoff_rate**(0.66667*self._b)*node_A**self.Qs_power_onAthresh
+
+        transport_capacity_prefactor_withA = self.Qs_prefactor*runoff_rate**(0.6+self._b/15.)*node_A**self.Qs_power_onA
+
+        internal_t = 0.
+        break_flag = False
+        dt_secs = dt*31557600.
+        counter = 0
+
+        while 1: #use the break flag, to improve computational efficiency for runs which are very stable
+            #we assume the drainage structure is forbidden to change during the whole dt
+            #print "loop..."
+            #note slopes will be *negative* at pits
+            #track how many loops we perform:
+            counter += 1
+            downward_slopes = node_S.clip(0.)
+            #positive_slopes = np.greater(downward_slopes, 0.)
+            transport_capacities_S = transport_capacity_prefactor_withA*(downward_slopes)**0.7
+            trp_diff = (transport_capacities_S - transport_capacities_thresh).clip(0.)
+            transport_capacities = np.sqrt(trp_diff*trp_diff*trp_diff)
+
+            if stability_condition == 'tight':
+                mock_diffusivities = np.zeros_like(transport_capacities, dtype=float)
+                mock_diffusivities = transport_capacities/downward_slopes
+                tstep_each_node = 10.*square_link_length/mock_diffusivities #we're relaxing the condition fivefold here, as the true VonNeumann condition is VERY restrictive
+                #if no node exceeds crit, tstep_each_node will just be nans and infs
+                delta_t_internal = np.nanmin(tstep_each_node) #in seconds, nanmin avoids the pit nodes
+                if delta_t_internal == np.inf: #no node exceeds crit
+                    delta_t_internal = dt_secs #nothing happened, so let the loop complete, awaiting more uplift
+                if internal_t + delta_t_internal >= dt_secs:
+                    dt_this_step = dt_secs-internal_t #now in seconds
+                    break_flag = True
+                else:
+                    dt_this_step = delta_t_internal #a min tstep was found (seconds). We terminate the loop
+            else: #loose, gradient based method
+                dt_this_step = dt_secs-internal_t #and the adjustment is made AFTER the dz calc
+
+            sed_into_node = np.zeros(grid.number_of_nodes, dtype=float)
+            dz = np.zeros(grid.number_of_nodes, dtype=float)
+            len_s_in = s_in.size
+            cell_areas = self.cell_areas
+
+            for i in s_in[::-1]: #work downstream
+                sed_flux_into_this_node = sed_into_node[i]
+                sed_flux_out_of_this_node = transport_capacities[i] #we work in volume flux, not volume per se here
+                flux_excess = sed_flux_into_this_node - sed_flux_out_of_this_node #gets deposited
+                dz[i] = flux_excess/cell_areas*dt_this_step
+                sed_into_node[flow_receiver[i]] += sed_flux_out_of_this_node
+
+            if stability_condition == 'loose':
+                elev_diff = node_z - node_z[flow_receiver]
+                delta_dz = dz[flow_receiver] - dz
+                node_flattening = self.fraction_gradient_change*elev_diff - delta_dz #note the condition is that gradient may not change by >X%, not must be >0
+                #note all these things are zero for a pit node
+                most_flattened_nodes = np.argmin(node_flattening[grid.core_nodes])
+                most_flattened_nodes = np.take(grid.core_nodes, most_flattened_nodes) #get it back to node number, not core_node number
+                most_flattened_val = np.take(node_flattening, most_flattened_nodes)
+                if most_flattened_val>=0.:
+                    break_flag = True #all nodes are stable
+                else: # a fraction < 1
+                    dt_fraction = self.fraction_gradient_change*np.take(elev_diff, most_flattened_nodes)/np.take(delta_dz, most_flattened_nodes)
+                    #print dt_fraction
+                    #correct those elevs
+                    dz *= dt_fraction
+                    dt_this_step *= dt_fraction
+
+            #print np.amax(dz), np.amin(dz)
+
+            node_z[grid.core_nodes] += dz[grid.core_nodes]
+
+            if break_flag:
+                break
+            #do we need to reroute the flow/recalc the slopes here? -> NO, slope is such a minor component of Diff we'll be OK
+            #BUT could be important not for the stability, but for the actual calc. So YES.
+            node_S = np.zeros_like(node_S)
+            #print link_length[core_draining_nodes]
+            node_S[core_draining_nodes] = (node_z-node_z[flow_receiver])[core_draining_nodes]/link_length[core_draining_nodes]
+            internal_t += dt_this_step #still in seconds, remember
+
+        self.grid=grid
+
         active_nodes = np.where(grid.status_at_node != CLOSED_BOUNDARY)[0]
         if io:
             try:
-                io[active_nodes] += dz[active_nodes]
+                io[active_nodes] += node_z[active_nodes]
             except TypeError:
                 if type(io)==str:
                     elev_name = io
             else:
-                return grid, io, capacity
+                return grid, io
 
         else:
-            elev_name = 'topographic__elevation'
+            elev_name = node_elevs
 
-        grid.at_node[elev_name][active_nodes] += dz[active_nodes]
-        grid.at_node['sediment_flux_capacity'] = capacity
+        if self.return_ch_props:
+            #add the channel property field entries,
+            #'channel_width', 'channel_depth', and 'channel_discharge'
+            Q = self.k_Q*runoff_rate*node_A**self._c
+            W = self.k_w*Q**self._b
+            H = Q**(0.6*(1.-self._b))*(self.mannings_n/self.k_w)**0.6*node_S**-0.3
+            tau = self.fluid_density*self.g*H*node_S
+            grid.at_node['channel_width'] = W
+            grid.at_node['channel_depth'] = H
+            grid.at_node['channel_discharge'] = Q
+            grid.at_node['channel_bed_shear_stress'] = tau
 
-        return grid, grid.at_node[elev_name], capacity
+
+        grid.at_node['fluvial_sediment_transport_capacity'] = transport_capacities
+        grid.at_node['fluvial_sediment_flux_into_node'] = sed_into_node
+        #elevs set automatically to the name used in the function call.
+        if stability_condition == 'tight':
+            grid.at_node['effective_fluvial_diffusivity'] = mock_diffusivities
+        self.iterations_in_dt = counter
+
+        return grid, grid.at_node[elev_name]
